Update HelperImageUploader.php
You can't cast it to INT because of large numbers. This number can be a double, and so can't be casted into INT number. The result is a negative number.
You can try with a large number such a post_max_size & upload_max_filesize : 8388608000. Product image won't work because of this bug.